add const to vectors passed as references in methods but not intended to be modified

+++ niryo_robot_hardware_stack/dynamixel_driver/src/single_motor_cmd.cpp
@@ -1,4 +1,5 @@
 #include "dynamixel_driver/single_motor_cmd.hpp"
+#include <sstream>
 
 namespace DynamixelDriver
 {
@@ -51,7 +52,7 @@ namespace DynamixelDriver
                 ss << "Learning mode";
                 break;
             default:
-                ss << "Unknown type " << _type;
+                ss << "Unknown type " << static_cast<int>(_type);
             break;
         }
